AudioLinkControllerLogo: added audio reactivity effect to logo, added AudioLink logotype, adjusted position/scale
This includes a new audio reactive shader matched with the existing ALControllerLogo.fbx model containing the mesh named "AudioLinkLogo". The fbx file was updated with the logotype merged into the existing mesh. As well, the placement, scale, and orientation of the complete logo were updated to accommodate the additional geo.